refactor(api): extract status endpoint from routes.ts
Move GET /api/status to its own file, leaving routes.ts as a pure
route registry (addRoute/matchRoute).
