+++ src/api/server.ts
@@ -27,6 +27,7 @@ import "./cron.js";
 import "./config.js";
 import "./logs.js";
 import "./usage.js";
+import "./status.js";
 import "./sessions.js";
 import "./subagents.js";
